second commit

+++ SPO2/Algo.h
@@ -1,6 +1,6 @@
 #pragma once
 
-extern void SelectionSort(int arr[], long size);
+extern void SelectionSort(int arr[], int size);
 
 extern void MergeSort(int* pArr, int bInd, int eInd);
